% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equivalencia_cno}
\alias{equivalencia_cno}
\title{Tabla de equivalencias CIUO-2008 <-> CNO 2015}
\format{
Un \code{data.frame} con columnas:
\describe{
\item{cod_ciuo_2008}{Codigo CIUO-2008 (4 digitos).}
\item{desc_ciuo_2008}{Descripcion del grupo primario CIUO-2008.}
\item{cod_cno_2015}{Codigo CNO 2015 equivalente (4 digitos).}
\item{desc_cno_2015}{Descripcion del grupo primario CNO 2015.}
\item{enlace_ciuo_cno}{Tipo de enlace de CIUO hacia CNO.}
\item{enlace_cno_ciuo}{Tipo de enlace de CNO hacia CIUO.}
}
}
\source{
INEI - Tabla de equivalencias CIUO-2008 / CNO 2015.
}
\usage{
data(equivalencia_cno)
}
\description{
Tabla de correspondencias entre los codigos CIUO-2008 y el Clasificador
Nacional de Ocupaciones 2015 (CNO 2015, 4 digitos). Publicada por el INEI
para facilitar la homologacion de series temporales entre clasificadores.
}
\keyword{datasets}
